requirements.txt
pinecone-client==3.0.1
openai==0.28.0
python-dotenv==1.0.0
numpy>=1.26.0